1654 랜선 자르기 통과

+++ BAEKJOON/1000~/1654_랜선 자르기/Source.cpp
@@ -27,7 +27,7 @@ int main()
 	}
 		
 
-	int nStart, nEnd, nFind, nResult;
+	unsigned long long nStart, nEnd, nFind, nResult;
 	nStart = 1;
 	nEnd = nMax;
 	nResult = 0;
